+++ Components/HeaderFooter/headerfooter.css
@@ -1,0 +1,54 @@
+header {
+  top: 0;
+  box-shadow: 0 5px 20px black;
+  background-color: var(--nav-color);
+  justify-content: space-around;
+  color: white;
+}
+
+.front {
+  border: 1px solid black;
+  border-radius: 20px;
+  left: 50%;
+  transform: translate(-50%);
+  display: flex;
+
+  align-items: center;
+  position: fixed;
+  z-index: 1;
+  width: 90vw;
+  height: 80px;
+}
+footer {
+  justify-content: center;
+  bottom: 0;
+  box-shadow: 0 -5px 20px black;
+  background-color: var(--nav-color);
+}
+
+.nav__item {
+  height: 78px;
+  width: 80px;
+  padding-top: 5px;
+  text-align: center;
+  border-radius: 10px;
+  transition: background-color 0.2s linear;
+}
+.nav__item:hover {
+  background-color: var(--hover-color);
+  cursor: pointer;
+}
+
+.nav__item-color {
+  background-color: var(--nav-item-color);
+}
+@media only screen and (min-width: 600px) {
+  .front {
+    width: 540px;
+  }
+  .nav__item {
+    margin-left: 10px;
+    margin-right: 10px;
+    width: 100px;
+  }
+}
