--- components/card.css
@@ -1,111 +1,0 @@
-.cardcontainer {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-.card {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  justify-content: center;
-  width: 90ve;
-  margin: 30px;
-  background-color: var(--card-color);
-  box-shadow: 0 0 20px black;
-  border-radius: 20px;
-  position: relative;
-
-  transition: opacity 1s linear;
-  opacity: 100%;
-}
-
-.card__question {
-  text-align: center;
-  margin-bottom: 20px;
-  margin-top: 40px;
-}
-.card__showanswer {
-  padding: 5px;
-  margin-top: 20px;
-  margin-bottom: 30px;
-  background-color: var(--card-show-answer-color);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid black;
-  border-radius: 10px;
-  transition: background-color 0.2s linear;
-}
-.card__answer {
-  text-align: center;
-  margin-bottom: 20px;
-  font-weight: bolder;
-
-  background-color: var(--answer-color);
-}
-
-.card__showanswer:hover {
-  color: black;
-  background-color: var(--hover-color);
-  cursor: pointer;
-}
-.card__bookmark {
-  position: absolute;
-  height: 60px;
-  width: 60px;
-  right: 20px;
-  top: -20px;
-  transition: 0.2s linear;
-}
-.card__bookmark:hover {
-  cursor: pointer;
-  content: url(../images/bookmark_black.svg);
-}
-
-.card__bookmark-big {
-  height: 80px;
-  width: 80px;
-  right: 10px;
-  top: -30px;
-}
-
-.card__tags {
-  width: 100%;
-  margin-left: 20px;
-  margin-bottom: 10px;
-  justify-content: left;
-  display: flex;
-  font-size: 20px;
-}
-
-.card__tag {
-  border: 1px solid black;
-  border-radius: 10px;
-  padding: 5px;
-  margin: 5px;
-  background-color: var(--card-color);
-  list-style-type: none;
-}
-.card__bigtext {
-  font-size: 26px;
-  padding-left: 10px;
-  padding-right: 10px;
-}
-
-.card__answer.card__bigtext {
-  font-size: 32px;
-  padding: 10px;
-}
-
-@media only screen and (min-width: 600px) {
-  .card {
-    width: 500px;
-  }
-}
-.display-none {
-  display: none;
-}
-
-.card-dissapear {
-  opacity: 0%;
-}
