--- components/form.css
@@ -1,38 +1,0 @@
-.input {
-  border-radius: 10px;
-  border: 2px solid black;
-  box-shadow: 5px 5px 10px black;
-  font-size: 20px;
-  margin-bottom: 20px;
-  background-color: var(--card-color);
-  width: 100%;
-  padding: 10px;
-}
-.form {
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  width: 100%;
-}
-.form__label {
-  font-size: 26px;
-}
-
-button.input {
-  transition: background-color 0ss linear;
-}
-
-button.input:hover {
-  background-color: var(--hover-color);
-  cursor: pointer;
-  color: black;
-}
-
-.randombutton:disabled {
-  background-color: grey;
-}
-
-.counter {
-  text-align: right;
-  margin-right: 10px;
-}
