--- components/profile.css
@@ -1,57 +1,0 @@
-.profile {
-  padding-top: 20px;
-  padding-bottom: 20px;
-}
-.profile__user {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-}
-.profile__counter {
-  width: 80px;
-  background-color: var(--card-color);
-  border: 1px solid black;
-  border-radius: 10px;
-  box-shadow: 5px 5px 10px black;
-}
-
-.profile__switch {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-.profile__text {
-  font-size: 30px;
-  font-weight: bold;
-  padding: 0 15px;
-}
-
-.profile__text-2 {
-  margin-bottom: 20px;
-}
-.profile__switch-image {
-  transition: 0.2s linear;
-}
-.profile__switch-image:hover {
-  cursor: pointer;
-  width: 80px;
-  height: 80px;
-}
-
-.profile__language-buttons {
-  font-size: 22px;
-  font-weight: bold;
-  border: 3px solid black;
-  border-radius: 10px;
-  padding: 5px;
-  background-color: var(--card-color);
-  padding: 5px 10px;
-  transition: background-color 0.2s linear;
-}
-
-.profile__language-buttons:hover {
-  background-color: var(--hover-color);
-  cursor: pointer;
-  color: black;
-}
