+++ form.js
@@ -6,10 +6,13 @@ import {
   tagsNew,
   bookmarkedNew,
   isAnswerShownNew,
-} from "./startingCardContent.js";
-import { addNewCardToVariables, initSubmitButton } from "./createCard.js";
-
-const htmlDecode = require("html-entities");
+} from "./Utils/startingCardContent.js";
+import {
+  addNewCardToVariables,
+  initSubmitButton,
+} from "./Components/Card/createCard.js";
+import { getQuestion } from "./Components/Form/fetch.js";
+import { changeFormTheme } from "./Components/Theme/theme.js";
 
 const webElements = {
   form: document.querySelector('[data-js="form"]'),
@@ -23,7 +26,7 @@ const webElements = {
   buttomRandom: document.querySelector('[data-js="button_random"]'),
   randomOutput: document.querySelector('[data-js="randomOutput"]'),
 };
-
+/* 
 let {
   form,
   main,
@@ -35,28 +38,23 @@ let {
   inputTag3,
   buttomRandom,
   randomOutput,
-} = webElements;
+} = webElements; */
 
 const cardSuite = {
   questions: questionsNew,
-
   answers: answersNew,
-
   questionsGerman: questionsGermanNew,
-
   answersGerman: answersGermanNew,
-
   tags: tagsNew,
-
   bookmarked: bookmarkedNew,
 };
 let { questions, answers, questionsGerman, answersGerman, tags, bookmarked } =
   cardSuite;
 
-form.addEventListener("submit", (event) => {
+webElements.form.addEventListener("submit", (event) => {
   event.preventDefault();
 
-  const formData = new FormData(event.target);
+  /* const formData = new FormData(event.target);
   const dataObject = Object.fromEntries(formData);
 
   const section = document.createElement("section");
@@ -111,7 +109,7 @@ form.addEventListener("submit", (event) => {
     answer.classList.toggle("card--dark-mode");
     showanswer.classList.toggle("card--dark-mode");
     section.classList.toggle("card--dark-mode");
-  }
+  } */
 
   getSaved();
   addNewCardToVariables(cardSuite, webElements);
@@ -132,7 +130,7 @@ function getSaved() {
     bookmarked = JSON.parse(sessionStorage.getItem("bookmarked"));
   }
 }
-addNewCardToVariables(cardSuite, webElements);
+//addNewCardToVariables(cardSuite, webElements);
 
 function saveCard() {
   sessionStorage.setItem("questions", JSON.stringify(questions));
@@ -143,7 +141,7 @@ function saveCard() {
   sessionStorage.setItem("bookmarked", JSON.stringify(bookmarked));
 }
 
-function addTagToList(inputTag, tagList) {
+/* function addTagToList(inputTag, tagList) {
   const tag = document.createElement("li");
   tag.classList.add("card__tag", "shadow");
   tag.textContent = "#" + inputTag.value;
@@ -151,76 +149,26 @@ function addTagToList(inputTag, tagList) {
     tag.classList.toggle("card--dark-mode");
   }
   tagList.append(tag);
-}
+} */
 
 const counterQuestion = document.querySelector('[data-js="question-counter"]');
 const counterAnswer = document.querySelector('[data-js="answer-counter"]');
 
-inputQuestion.addEventListener("input", (event) => {
+webElements.inputQuestion.addEventListener("input", (event) => {
   const lettersleft = 150 - Number(event.target.value.length);
   counterQuestion.textContent = lettersleft + " characters left.";
 });
-inputAnswer.addEventListener("input", (event) => {
+webElements.inputAnswer.addEventListener("input", (event) => {
   const lettersleft = 150 - Number(event.target.value.length);
   counterAnswer.textContent = lettersleft + " characters left.";
 });
 
 let darkmodeOn = sessionStorage.getItem("darkModeOn");
 
-function changeTheme() {
-  body.classList.toggle("body--dark-mode");
-  const navItemSelected = document.querySelector(
-    '[data-js="nav__item-selected"]'
-  );
-  const header = document.querySelector('[data-js="header"]');
-  const footer = document.querySelector('[data-js="footer"]');
-  navItemSelected.classList.toggle("nav--dark-mode");
-  form.classList.toggle("card--dark-mode");
-  inputQuestion.classList.toggle("card--dark-mode");
-  inputAnswer.classList.toggle("card--dark-mode");
-  inputTag1.classList.toggle("card--dark-mode");
-  inputTag2.classList.toggle("card--dark-mode");
-  inputTag3.classList.toggle("card--dark-mode");
-  button = document.querySelector('[data-js="button"]');
-  button.classList.toggle("card--dark-mode");
-  buttomRandom.classList.toggle("card--dark-mode");
-  randomOutput.classList.toggle("card--dark-mode");
-  header.classList.toggle("card--dark-mode");
-  footer.classList.toggle("card--dark-mode");
-}
 if (darkmodeOn == "1") {
-  changeTheme();
+  changeFormTheme(webElements);
 }
 
-buttomRandom.addEventListener("click", () => {
-  getQuestion();
+webElements.buttomRandom.addEventListener("click", () => {
+  getQuestion(webElements);
 });
-
-async function getQuestion() {
-  try {
-    buttomRandom.disabled = true;
-    buttomRandom.textContent = "...";
-    randomOutput.textContent = "Fetching Random Card";
-    const fetchJson = await fetch("https://opentdb.com/api.php?amount=1");
-    if (fetchJson.ok) {
-      const fetchJs = await fetchJson.json();
-
-      const question = htmlDecode.decode(fetchJs.results[0].question);
-      const answer = htmlDecode.decode(fetchJs.results[0].correct_answer);
-      const tag = htmlDecode.decode(fetchJs.results[0].category);
-
-      inputQuestion.value = question;
-      inputAnswer.value = answer;
-      inputTag1.value = tag;
-
-      randomOutput.textContent = "Fetching Successful";
-    }
-  } catch (error) {
-    console.log("Fetching failed...");
-    console.log(error);
-    randomOutput.textContent = "Fetching failed";
-  } finally {
-    buttomRandom.disabled = false;
-    buttomRandom.textContent = "Generate Random Card";
-  }
-}
