--- startingCardContent.js
@@ -1,41 +1,0 @@
-export const questionsNew = [
-  "question 1?",
-  "question 2?",
-  "question 3?",
-  "question 4?",
-  "question 5?",
-];
-
-export const answersNew = [
-  "answer 1.",
-  "answer 2.",
-  "answer 3",
-  "answer 4",
-  "answer 5",
-];
-
-export const questionsGermanNew = [
-  "Frage 1?",
-  "Frage 2?",
-  "Frage 3?",
-  "Frage 4?",
-  "Frage 5?",
-];
-
-export const answersGermanNew = [
-  "Antwort 1.",
-  "Antwort 2.",
-  "Antwort 3.",
-  "Antwort 4.",
-  "Antwort 5.",
-];
-
-export const tagsNew = [
-  ["tag1", "tag2", "tag3"],
-  ["tag1"],
-  ["tag1", "tag2"],
-  ["tag1", "tag2", "tag3"],
-  ["tag1", "tag2"],
-];
-export const bookmarkedNew = [false, false, false, false, false];
-export const isAnswerShownNew = [];
